+++ src/test/java/nextstep/subway/utils/apiHelper/LineSectionApiHelper.java
@@ -1,4 +1,4 @@
-package nextstep.subway.utils;
+package nextstep.subway.utils.apiHelper;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
